Refactor code structure
Move common code into a Utilities module and refactor the experiments
to be in a single file.

For the data generation, use witness files and integrate them
into the Makefile. Make the code multi-threaded.

This is a multi-step process and for now the "new" code will be in
src/ until the migration is complete.

Signed-off-by: Laslo Hunhold <hunhold@uni-koeln.de>