Refactor Makefile
Use suffix rules to express the transformation of .jl to .format and
.output. Properly handle errors.

This change treats the formatting of each source file as an atomic
operation. If make, thus, is executed with multiple threads it will
run the data generation and formatting in parallel, and it will only
format files that have been changed since the last formatting.

This change also allows formatting any arbitrary file (e.g.
"src/foo.jl") via

	make src/foo.format

even if the file is not mentioned in the Makefile.

Signed-off-by: Laslo Hunhold <hunhold@uni-koeln.de>