Add test matrix file generator
The MatrixDepot.jl package is really annoying to work with. Its
interface promises a lot of functionality, but it is broken insofar that
many of the referenced matrices have incomplete metadata. It's also very
chatty and slow and requires a constant internet connection. In the long
term, it's not a great solution for reproducibility. What if the sparse
matrix collection goes offline, or what if we don't have a constant
internet connection?

This is unacceptable, which is why this commit adds a way to prepare
a compressed binary blob of all tested matrices, in this case all
matrices from the Suite Sparse Matrix collection with NNZ < 50.000.

The resulting blob is merely 27MB in size and contains all the metadata
we need for our experiments without having to generate it at runtime
every time. In total we obtain roughly over 700 matrices, after
filtering out all matrices which have no full rank.

Signed-off-by: Laslo Hunhold <hunhold@uni-koeln.de>