Reduce the test matrix type complexity
The MatrixDepot properly uses Julia convention by expressing the
numerical properties of the matrices as provisioned for by the type
system, e.g. when a matrix is symmetric it is wrapped using Symmetric().
As far as I can tell it is only used to denote symmetric matrices.

This approach, while honourable, has many shortcomings. Most
importantly, given sparse matrices are generally poorly supported in
Julia, it leads to many cases where basic functionality is broken: You
cannot, for instance, simply call nnz() on such a matrix. Of course this
can be solved by implementing these interfaces, but we can't waste our
time on this right now. Additionally, it is inconsistent with other
matrix properties that are not exposed in the type system, e.g. positive
definiteness, etc.

Maybe there's an obvious way to work with these matrices other than
casting the matrix via parent() every time we want to actually do
something with it, but now it's become too much of a hassle! I am sick
and tired of Julia vomiting a 20-layer stack trace on me for the 1000th
time because yet another elementary function doesn't know how to deal
with the ingenious symmetric-wrapper-type.

We explicitly define the matrix in the TestMatrix struct to be a
Float64 sparse matrix and add a separate field to store the symmetric
property. While at it, we also add a nnz field.

Signed-off-by: Laslo Hunhold <hunhold@uni-koeln.de>