Move the generation of test matrices into a separate module
It turns out that Suppressor isn't able to shut MatrixDepot up when
the latter is imported in a module that is being imported, despite being
wrapped in a suppress_err-macro.

While having everything in one module TestMatrices is the most
desirable, it's also a good feeling to have significantly reduced the
surface area of modules that explicitly or implicitly import
MatrixDepot, and when the binary blobs are present, we don't even import
it at all.

Signed-off-by: Laslo Hunhold <hunhold@uni-koeln.de>