Heavily refactor Utilities into Experiments module
We will have a diverse set of experiments in the future, but they all
boil down to a single type and a single matrix being evaluated in some
form returning a measurement.

Making use of the type machinery in Julia, we define an Experiment type
that is a struct containing, besides the types and matrices to be
tested, a parameter of an abstract experiment parameter type.

We define, for each experiment type (in this case for the solver),
a type as a subtype of experiment parameter type to contain all the
solver-specific stuff, and also a dedicated type for the storage of
a single measurement.

While the machinery in Experiment does all the work, we only need to
define the function for obtaining a measurement. For each field in the
measurement struct, we automatically create the corresponding result
files, offering maximum flexibility.

Signed-off-by: Laslo Hunhold <hunhold@uni-koeln.de>