Properly handle preliminary problems and print experiment progress
By allowing the results to be nothing we have more flexibility of
catching errors early on without having to pass invalid matrices (e.g.
ones containing NaN or Inf, which we now check for) to the lower-level
measurement functions.

Additionally, add a way to print the progress of all threads by showing
the number of completed measurements and a list of currently processing
ones.

Signed-off-by: Laslo Hunhold <hunhold@uni-koeln.de>