Use SuiteSparse's SPQR for generating fill-reducing permutations for QR
Interestingly, this approach mostly yields upper triangular matrices
that are even more sparse than those returned by SPQR itself.

It seems, thus, that the most important aspect of QR is the generation
of the fill-reducing permutations. Even though we cannot directly
use SPQR given it requires Float32 or Float64, this is the first
reproduction of state-of-the-art QR that I know of.

Signed-off-by: Laslo Hunhold <hunhold@uni-koeln.de>