Refactor LU decomposer to use UMFPACK and add solve_lu.jl
To increase the meaningfulness of the results it makes most sense to
mirror state of the art methods as much as possible. For this reason,
the ideal case would be to be able to use UMFPACK as much as possible.

However, given that UMFPACK only supports Float32 and Float64, we cannot
do that. What we can do is run an LU decomposition using UMFPACK on
the Float64 version of a given matrix and simply use the row and
column permutations, directly apply them to the matrix in the specified
machine number type and perform an LU decomposition without
pivotisation.

This results in matrices that are generally identical to those returned by
UMFPACK.

Signed-off-by: Laslo Hunhold <hunhold@uni-koeln.de>