Add a module for converting between Float128 and other number formats
Using Float64 as a reference doesn't cut it, as Takum64 is superior in
precision. We can't, however, use Takum64 instead, as this would shadow
computational errors in a way.

The solution is to use the Quadmath package and use Float128 as a
reference. This requires converting between the different formats, which
is not difficult for BFloat16 and Posits, which can be losslessly
expanded to Float64 and then converted to Float128 using the provided
functions in Quadmath, but Takums require more work, given Takum64 is
more precise than Float64.

We end up with an almost complete codec, but it does the job.

Signed-off-by: Laslo Hunhold <hunhold@uni-koeln.de>