Add more error metrics and fix bug in output file generation
We add relative error and logarithmic relative error by Gustafson.
The output directory is restructured to contain the respective CSV
files in subdirectories.

We fix a small bug in the output file generation where we erroneously
call return in the loop, only generating the file for the first struct
entry.

The Makefile is amended to automatically remove the (empty) output
folders for each experiment.

Signed-off-by: Laslo Hunhold <hunhold@uni-koeln.de>