Transpose the output CSV files and detect type cast under-/overflows
It makes more sense to have each type's result in column form. We also
detect if a type is not able to represent the matrix within its dynamic
range, exiting early and treating it as a hard error (just like a
singular matrix).

The reasoning behind this is that while saturation arithmetic has its
merits, it should not be an unfair advantage over the normal
floating-point types which overflow to infinity.

Signed-off-by: Laslo Hunhold <hunhold@uni-koeln.de>