Pre-generate the row and column permutations
This saves a lot of effort, because we would previously generate
the permutations for each type even though the permutations are
independent of the type.

We subsequently change the qr() and lu() functions to directly take
the permuted matrix, simplifying them in the process.

Overall, this leads to only a minimal speedup versus the overhead we
have with out hand-rolled QR and LU routines, but it was sensible to do
it like this. Additionally, this reduces memory throughput.

Signed-off-by: Laslo Hunhold <hunhold@uni-koeln.de>