Add CSV-sorting postprocessor
One of the best ways to compare a wide range of experiment samples
is to just sort them and print them. This is comparable to the modern
approach of considering the empirical cumulative distribution function
(eCDF) for random variables instead of box plots, which can be
misleading.

The challenge herein lies in the fact that TikZ/PGF does not like zeros
or infinities for logarithmic plots. Thus we have to replace these
values with fitting finite values, which we then can subsequently
plot with special labels in TikZ/PGF. To have proper formatting, we
do an automatic fitment to the decadic logarithm scale.

Signed-off-by: Laslo Hunhold <hunhold@uni-koeln.de>