Add graph scraper
We use the big collection of graphs from https://networkrepository.com.
Given they do not provide a public API, we scrape the networks page for
all download links, reformat them, automatically check and download
them, extracting the contained .mtx or .edges files. Some special cases
needed to be covered due to inconsistencies in the data.

An efficient method is used to check if the respective graph is
too large for our liking (only a single HTTP request). In the end,
the uncompressed graph repository is around 700 MB.

The next stage is to generate the test matrices from them. We do that
by making use of out/graph_files, which is a list of all graph files.
A challenge will be to handle both mtx and edgelists seamlessly.

Signed-off-by: Laslo Hunhold <hunhold@uni-koeln.de>