Make eigenvector matching more robust
When eigenvalues are close to each other, the order of the eigenvectors
might be mixed up, yielding incorrect results. We now make use of cosine
similarity and the Hungarian algorithm to find the best permutation
such that the overall cosine similarity is maximised between the
reference solution and computed approximate eigenvectors.

Additionally, for the sign normalisation, we do not use the first
respective entry, as this might be close to zero, but instead the entry
index which turned out to be absolutely largest in the reference solution.

Signed-off-by: Laslo Hunhold <hunhold@uni-koeln.de>