Add an eigenvalue buffer zone
Given eigenvalues can be very close to each other, and the Krylov
methods do not guarantee the ordering across different arithmetics,
the case might happen that we get a cutoff between two very close
eigenvalues and obtain exactly the wrong eigenvector in our set.

We circumvent this by computing more eigenvalues/eigenvectors than
necessary, doing the cosine similarity matching on the full set and
then discarding the rest. This buffer method is a heuristic, which
only fails if we assume that the aforementioned case can happen across
5 eigenvalues, which would be absurd.

Signed-off-by: Laslo Hunhold <hunhold@uni-koeln.de>